refactor: modularize SuperUserSeedExtension for better structure
