+++ extensions/seed/super-user/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -19,4 +19,4 @@
 #
 #
 
-org.eclipse.edc.identityhub.demo.IdentityHubExtension
+org.eclipse.tractusx.identityhub.seed.SuperUserSeedExtension
