+++ runtimes/identityhub/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -20,4 +20,3 @@
 #
 
 org.eclipse.edc.identityhub.demo.IdentityHubExtension
-org.eclipse.edc.identityhub.seed.SuperUserSeedExtension
